Fix query test for charge
